+++ core/io/packet_peer.h
@@ -37,7 +37,6 @@
 
 #include "core/extension/ext_wrappers.gen.inc"
 #include "core/object/gdvirtual.gen.inc"
-#include "core/object/script_language.h"
 #include "core/variant/native_ptr.h"
 
 class PacketPeer : public RefCounted {
