+++ core/object/script_instance.cpp
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  webrtc_peer_connection_extension.h                                    */
+/*  script_instance.cpp                                                   */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -28,36 +28,44 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef WEBRTC_PEER_CONNECTION_EXTENSION_H
-#define WEBRTC_PEER_CONNECTION_EXTENSION_H
+#include "script_instance.h"
 
-#include "webrtc_peer_connection.h"
+#include "core/object/script_language.h"
 
-#include "core/extension/ext_wrappers.gen.inc"
-#include "core/object/gdvirtual.gen.inc"
-#include "core/variant/native_ptr.h"
+Variant ScriptInstance::call_const(const StringName &p_method, const Variant **p_args, int p_argcount, Callable::CallError &r_error) {
+	return callp(p_method, p_args, p_argcount, r_error);
+}
 
-class WebRTCPeerConnectionExtension : public WebRTCPeerConnection {
-	GDCLASS(WebRTCPeerConnectionExtension, WebRTCPeerConnection);
+void ScriptInstance::get_property_state(List<Pair<StringName, Variant>> &state) {
+	List<PropertyInfo> pinfo;
+	get_property_list(&pinfo);
+	for (const PropertyInfo &E : pinfo) {
+		if (E.usage & PROPERTY_USAGE_STORAGE) {
+			Pair<StringName, Variant> p;
+			p.first = E.name;
+			if (get(p.first, p.second)) {
+				state.push_back(p);
+			}
+		}
+	}
+}
 
-protected:
-	static void _bind_methods();
+void ScriptInstance::property_set_fallback(const StringName &, const Variant &, bool *r_valid) {
+	if (r_valid) {
+		*r_valid = false;
+	}
+}
 
-public:
-	/** GDExtension **/
-	EXBIND0RC(ConnectionState, get_connection_state);
-	EXBIND0RC(GatheringState, get_gathering_state);
-	EXBIND0RC(SignalingState, get_signaling_state);
-	EXBIND1R(Error, initialize, Dictionary);
-	EXBIND2R(Ref<WebRTCDataChannel>, create_data_channel, String, Dictionary);
-	EXBIND0R(Error, create_offer);
-	EXBIND2R(Error, set_remote_description, String, String);
-	EXBIND2R(Error, set_local_description, String, String);
-	EXBIND3R(Error, add_ice_candidate, String, int, String);
-	EXBIND0R(Error, poll);
-	EXBIND0(close);
+Variant ScriptInstance::property_get_fallback(const StringName &, bool *r_valid) {
+	if (r_valid) {
+		*r_valid = false;
+	}
+	return Variant();
+}
 
-	WebRTCPeerConnectionExtension() {}
-};
+const Variant ScriptInstance::get_rpc_config() const {
+	return get_script()->get_rpc_config();
+}
 
-#endif // WEBRTC_PEER_CONNECTION_EXTENSION_H
+ScriptInstance::~ScriptInstance() {
+}
