+++ core/variant/variant_parser.cpp
@@ -32,6 +32,7 @@
 
 #include "core/input/input_event.h"
 #include "core/io/resource_loader.h"
+#include "core/object/script_language.h"
 #include "core/os/keyboard.h"
 #include "core/string/string_buffer.h"
