+++ editor/editor_script.cpp
@@ -32,6 +32,7 @@
 
 #include "editor/editor_interface.h"
 #include "editor/editor_node.h"
+#include "scene/main/node.h"
 
 void EditorScript::add_root_node(Node *p_node) {
 	if (!EditorNode::get_singleton()) {
