+++ editor/editor_script.h
@@ -33,10 +33,10 @@
 
 #include "core/object/gdvirtual.gen.inc"
 #include "core/object/ref_counted.h"
-#include "core/object/script_language.h"
 
 class EditorInterface;
 class EditorNode;
+class Node;
 
 class EditorScript : public RefCounted {
 	GDCLASS(EditorScript, RefCounted);
