+++ editor/editor_translation_parser.h
@@ -34,7 +34,6 @@
 #include "core/error/error_list.h"
 #include "core/object/gdvirtual.gen.inc"
 #include "core/object/ref_counted.h"
-#include "core/object/script_language.h"
 #include "core/variant/typed_array.h"
 
 class EditorTranslationParserPlugin : public RefCounted {
