+++ editor/plugins/editor_preview_plugins.cpp
@@ -33,6 +33,7 @@
 #include "core/config/project_settings.h"
 #include "core/io/file_access_memory.h"
 #include "core/io/resource_loader.h"
+#include "core/object/script_language.h"
 #include "core/os/os.h"
 #include "editor/editor_paths.h"
 #include "editor/editor_scale.h"
