+++ editor/plugins/editor_resource_tooltip_plugins.h
@@ -33,7 +33,6 @@
 
 #include "core/object/gdvirtual.gen.inc"
 #include "core/object/ref_counted.h"
-#include "core/object/script_language.h"
 #include <scene/gui/control.h>
 
 class Control;
