+++ editor/plugins/packed_scene_translation_parser_plugin.cpp
@@ -31,6 +31,7 @@
 #include "packed_scene_translation_parser_plugin.h"
 
 #include "core/io/resource_loader.h"
+#include "core/object/script_language.h"
 #include "scene/gui/option_button.h"
 #include "scene/resources/packed_scene.h"
