+++ editor/rename_dialog.cpp
@@ -40,8 +40,8 @@
 #include "editor/editor_string_names.h"
 #include "editor/editor_themes.h"
 #include "editor/editor_undo_redo_manager.h"
+#include "editor/plugins/script_editor_plugin.h"
 #include "modules/regex/regex.h"
-#include "plugins/script_editor_plugin.h"
 #include "scene/gui/check_box.h"
 #include "scene/gui/check_button.h"
 #include "scene/gui/control.h"
