+++ modules/gdscript/gdscript_rpc_callable.cpp
@@ -30,6 +30,7 @@
 
 #include "gdscript_rpc_callable.h"
 
+#include "core/object/script_language.h"
 #include "core/templates/hashfuncs.h"
 #include "scene/main/node.h"
