+++ modules/mono/editor/code_completion.cpp
@@ -31,6 +31,7 @@
 #include "code_completion.h"
 
 #include "core/config/project_settings.h"
+#include "core/object/script_language.h"
 #include "editor/editor_file_system.h"
 #include "editor/editor_settings.h"
 #include "scene/gui/control.h"
