+++ modules/multiplayer/scene_cache_interface.h
@@ -33,6 +33,7 @@
 
 #include "scene/main/multiplayer_api.h"
 
+class Node;
 class SceneMultiplayer;
 
 class SceneCacheInterface : public RefCounted {
