+++ scene/resources/style_box.h
@@ -34,7 +34,6 @@
 #include "core/io/resource.h"
 #include "core/object/class_db.h"
 #include "core/object/gdvirtual.gen.inc"
-#include "core/object/script_language.h"
 
 class CanvasItem;
