+++ servers/audio/audio_effect.h
@@ -34,7 +34,6 @@
 #include "core/io/resource.h"
 #include "core/math/audio_frame.h"
 #include "core/object/gdvirtual.gen.inc"
-#include "core/object/script_language.h"
 #include "core/variant/native_ptr.h"
 
 class AudioEffectInstance : public RefCounted {
