+++ servers/text/text_server_extension.h
@@ -32,7 +32,6 @@
 #define TEXT_SERVER_EXTENSION_H
 
 #include "core/object/gdvirtual.gen.inc"
-#include "core/object/script_language.h"
 #include "core/os/thread_safe.h"
 #include "core/variant/native_ptr.h"
 #include "core/variant/typed_array.h"
